backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.1
python-multipart==0.0.9
timm==1.0.3
Pillow==10.3.0
pydantic==2.7.4
pydantic-settings==2.3.4
huggingface-hub==0.23.4
